refactor(admin): 收口赛事设置事实展示
